requirements.txt
discord.py
sqlalchemy
demoparser2
